Optimize by using hash function
I try to reduce the cache miss and the execution time of findName()
by using the hash table method. I choose BKDRHash as the hash function
to build the hash table. Therefore, the number of the cache miss is
decreased by 45%, and the execution time of findName() is almost
approaches to 0 second. This optimization has a side-effect that the
execution time of append() is increasing about 27% due to the function
becomes a bit more complicated.