Implement BST
I build a BST to substitute for the linked list as the data structure
in the hope of improving the searching performance. Therefore, the
outcome shows that the execution time of findName() is almost down
to 0, and the execution time of append() is increasing about 0.03 second
because building the tree takes time.