Modify node struct of BST
I modified the node struct of BST in order to fit the size of cache line.
With the hope of reducing cache-miss and the append() function's execution
time, the outcome shows that the number of cache-miss is decreased about
180,000 times, and the execution time of append() is also decreased about
0.006 second which is 17% of the execution time of the original BST's
append().

Main changes from the original BST:
Add 5 more phonebook_entry pointer to the node struct of BST, and some
functions are modified to keep this program working properly.